+++ CHANGES.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+2.0.7 (unreleased)
+------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
+
+
 2.0.6 (2016-08-18)
 ------------------